docs/requirements.txt
pyaml==15.8.2
ansible==2.1.1.0

